requirements.txt
streamlit
pyyaml
openai
spacy 